#36 extract removeItemsFromCart with slight modifications regarding async on callers
