+++ Project/Frontend/src/Order.jsx
@@ -10,7 +10,7 @@ import PaymentModal from './PaymentModal';
 import NewOrderModal from './NewOrderModal';
 import MenuCard from './MenuCard';
 import { Link } from 'react-router-dom';
-import { addItemToCart, replaceSuggestion, clearCart } from './CartContext/cartUtils';
+import { addItemToCart, replaceSuggestion, clearCart, removeItemFromCart } from './CartContext/cartUtils';
 
 // Popup component to display order information
 const OrderInfoPopup = React.memo(({
@@ -163,7 +163,7 @@ OrderInfoPopup.propTypes = {
 
 function Order() {
     // Context and state variables
-    const { cart, fetchCart, menuItems, removeItemFromCart, customer, submitOrder, createNewOrder, tableNum, setCart, setCustomer, setTableNum, suggestions, setSuggestions } = useContext(CartContext);
+    const { cart, fetchCart, menuItems, customer, submitOrder, createNewOrder, tableNum, setCart, setCustomer, setTableNum, suggestions, setSuggestions } = useContext(CartContext);
     const [message, setMessage] = useState(''); // snackbar message
     const [severity, setSeverity] = useState('success'); // snackbar severity
     const [showOrderInfo, setShowOrderInfo] = useState(() => { // Controls order info popup visibility
@@ -443,8 +443,11 @@ function Order() {
     }, [customer?.customerId]); // Only run when customerId changes
 
     // Function to decrease item quantity by removing it from the cart
-    const decreaseItemQuantity = (itemId) => {
-        removeItemFromCart(itemId, false);
+    const decreaseItemQuantity = async (itemId) => {
+        const result = await removeItemFromCart(customer, cart, itemId, false);
+        if (result){
+            fetchCart(customer).then(setCart);
+        }
     };
 
     // Function to increase item quantity by adding it to the cart
@@ -692,7 +695,12 @@ function Order() {
                                 </IconButton>
                             </Box>
                             <Button
-                                onClick={() => removeItemFromCart(item.itemId, true)}
+                                onClick={async () => {
+                                    const result = await removeItemFromCart(customer, cart, item.itemId, true);
+                                           if (result){
+                                                fetchCart(customer).then(setCart);
+                                            }
+                                }}
                                 sx={{ backgroundColor: '#333', color: 'white', '&:hover': { backgroundColor: 'darkgray' } }}
                             >
                                 Remove All
